Refactor backend code into separate files and use Google Maps element for route drawing
