+++ admin.html
@@ -37,7 +37,7 @@
           </section>
           <section>
             <!-- Routes -->
-            <tracking-routes></tracking-routes>
+            <tracking-routes lat="42.730172" lng="-73.678803" zoom="15"></tracking-routes>
           </section>
           <section>
             <!-- Stops -->
