+++ bower.json
@@ -12,6 +12,7 @@
     "components"
   ],
   "dependencies": {
+    "google-map": "GoogleWebComponents/google-map#^1.0.0",
     "iron-elements": "PolymerElements/iron-elements#~1.0.3",
     "paper-elements": "PolymerElements/paper-elements#~1.0.5",
     "polymer": "Polymer/polymer#^1.1.0"
