+++ static/components/tracking-vehicles.html
@@ -1,9 +1,29 @@
 <link rel="import" href="../../bower_components/iron-form/iron-form.html">
+<link rel="import" href="../../bower_components/iron-icon/iron-icon.html">
+<link rel="import" href="../../bower_components/paper-card/paper-card.html">
 <link rel="import" href="../../bower_components/paper-input/paper-input.html">
 <link rel="import" href="../../bower_components/paper-button/paper-button.html">
-<link rel="import" href="../../bower_components/iron-validatable-behavior/iron-validatable-behavior.html">
 
 <dom-module id="tracking-vehicles">
+  <style>
+    .vehicle-form {
+      margin: 25px;
+    }
+    .vehicle-card {
+      margin: 5px;
+    }
+    .vehicle-list h1 {
+      margin-left: 15px;
+    }
+    .vehicle-list .vehicle-card {
+      margin-left: 15px;
+    }
+    .vehicle-action {
+      --iron-icon-height: 20px;
+      --iron-icon-width: 20px;
+    }
+  </style>
+
   <template>
     <iron-ajax
       auto
@@ -13,21 +33,38 @@
       last-response="{{vehicles}}">
     </iron-ajax>
 
-    <h1> Vehicles </h1>
-    <template is="dom-if" if="{{vehicles}}">
-      <template is="dom-repeat" items="{{vehicles}}" as="vehicle">
-        <b> Name: </b> <span>{{vehicle.vehicleName}}</span>
-        <b> ID: </b> <span>{{vehicle.vehicleID}}</span> 
-        <br />
+    <div class="vehicle-list">
+      <h1> Vehicles </h1>
+      <template is="dom-if" if="{{vehicles}}">
+        <template is="dom-repeat" items="{{vehicles}}" as="vehicle">
+          <paper-card class="vehicle-card" heading="{{vehicle.vehicleName}}">
+            <div class="card-content">
+              <b> ID: </b> <span>{{vehicle.vehicleID}}</span>
+            </div>
+            <div class="card-actions">
+              <paper-button>
+                <iron-icon class="vehicle-action" icon="history"></iron-icon>
+              </paper-button>
+              <paper-button> 
+                <iron-icon class="vehicle-action" icon="create"></iron-icon>
+              </paper-button>
+              <paper-button>
+                <iron-icon class="vehicle-action" icon="delete"></iron-icon>
+              </paper-button>
+            </div>
+          </paper-card>
+        </template>
       </template>
-    </template>
+    </div>
 
-    <h1> New Vehicle </h1>
-    <form is="iron-form" id="vehicleForm" content-type="application/json" method="post" action="/vehicles/create">
-      <paper-input name="vehicleName" label="Vehicle Name"></paper-input>
-      <paper-input name="vehicleID" label="Vehicle ID"></paper-input>
-      <paper-button raisedButton onclick="submitVehicle()">Submit</paper-button>
-    </form>
+    <div class="vehicle-form">
+      <h1> New Vehicle </h1>
+      <form is="iron-form" id="vehicleForm" content-type="application/json" method="post" action="/vehicles/create">
+        <paper-input name="vehicleName" label="Vehicle Name"></paper-input>
+        <paper-input name="vehicleID" label="Vehicle ID"></paper-input>
+        <paper-button raisedButton onclick="submitVehicle()">Submit</paper-button>
+      </form>
+    </div>
   </template>
 
   <script>
